% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{Plot mCPR data and estimates}
\usage{
plot_cp(dat, est, iso_code, CI = 95)
}
\arguments{
\item{dat}{tibble which contains mCPR observations. Columns: iso, year, cp}

\item{est}{tibble which contains mCPR estimates. Columns: “Country or area”, iso, Year, Median, U95,L95}

\item{iso_code}{Numeric. The ISO code of the country for which to generate the plot.}

\item{CI}{confidence intervals to be plotted. Options can be: 80, 95, or NA (no CI plotted)}
}
\value{
A ggplot object showing modern contraceptive use over time.
}
\description{
Plot mCPR data and estimates
}
\examples{
Load data and generate plot for Afghanistan (ISO code 4) with 95\% CI

plot_cp(dat, est, iso_code = 4, CI = 95)

# Generate plot for Kenya (ISO code 404) with 80\% CI
plot_cp(dat, est, iso_code = 404, CI = 80)

}
